Virtual garage door "simple" config
This allows you to use a mimo-lite relay to control your garage door and use the simple config. If Homebridge-Homeseer4 see's a virtual device type under the garage door, it will instead use "opening" and "closing" controls, in order to effectively show the garage door opening and closing in HomeKit. You will need to set up a virtual garage door device in HomeSeer, and corresponding events to control it. I also have a wireless sensor that shows when the garage door is fully open, and use the mimo-lite sensor to show when the door is fully closed. Virtual device/events setup screenshots: https://imgur.com/a/i1JHeGi